Move apis to common folder and fix paths
